Fixed a bunch of frontend stuff and messageing
