+++ backend/meeting/src/utils/index.js
@@ -10,7 +10,7 @@ const { ValidationError } = require("./error/app-errors");
 module.exports.ValidateSignature = async req => {
 	try {
 		const signature = req.get("Authorization");
-		console.log(signature);
+		// console.log(signature);
 		const payload = await jwt.verify(signature.split(" ")[1], APP_SECRET);
 		req.user = payload;
 		return true;
