+++ frontend/src/Components/LoginValidation.js
@@ -1,14 +1,14 @@
 import React, { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { ToastContainer, toast } from "react-toastify";
-// import { SERVER_URL } from "../config";
 import { useUpdateUserContext } from "../contexts/LoginContext";
 import APIHandler from "../utils/api-methods";
+import { useToastUpdate } from "../contexts/PageContext";
 
 //Component for Login validation
 const Login = () => {
-	const api = new APIHandler();
 	const navigate = useNavigate();
+	const { sendToastSuccess, sendToastError } = useToastUpdate();
+
 	const [inputValue, setInputValue] = useState({
 		email: "",
 		password: "",
@@ -25,21 +25,12 @@ const Login = () => {
 		});
 	};
 
-	const handleError = err =>
-		toast.error(err, {
-			position: "bottom-left",
-		});
-
-	const handleSuccess = msg =>
-		toast.success(msg, {
-			position: "bottom-right",
-		});
-
 	const handleSubmit = async e => {
 		e.preventDefault();
 
 		try {
 			// Send a POST request to login
+			const api = new APIHandler();
 			const { data } = await api.PostData(
 				"/user/login",
 				{ ...inputValue },
@@ -47,17 +38,19 @@ const Login = () => {
 			);
 			// Set the token in local storage
 			await setAuthToken(data.token);
-			handleSuccess(data.message);
+			sendToastSuccess(data.message);
 			saveUser(data.existingUser);
 			updateLoginStatus(true);
+
 			// Redirect to the "/profile" route after successful login
 			setTimeout(() => {
 				navigate("/profile");
 			}, 2000);
 		} catch (error) {
 			console.error(error);
-			handleError(error.response.data);
+			sendToastError(error.response.data);
 		}
+
 		// Clear the password input field
 		setInputValue({
 			...inputValue,
@@ -66,53 +59,48 @@ const Login = () => {
 	};
 
 	return (
-		<div className="form_container">
-			<form onSubmit={handleSubmit}>
-				<div>
-					<label htmlFor="email" className="input_label">
-						Enter your email
-						<b>*</b>
-					</label>
-				</div>
-				<input
-					className="input_margin"
-					name="email"
-					type="email"
-					placeholder="Email"
-					autoComplete="email"
-					value={email}
-					onChange={handleOnChange}
-				/>
+		<form onSubmit={handleSubmit} className="form_container">
+			<div>
+				<label htmlFor="email" className="input_label">
+					Enter your email
+					<b>*</b>
+				</label>
+			</div>
+			<input
+				className="input_margin"
+				name="email"
+				type="email"
+				placeholder="Email"
+				autoComplete="email"
+				value={email}
+				onChange={handleOnChange}
+			/>
 
-				<div>
-					<label htmlFor="password" className="input_label">
-						Enter your password
-						<b>*</b>
-					</label>
-				</div>
-				<input
-					className="input_margin"
-					name="password"
-					type="password"
-					placeholder="Password"
-					autoComplete="password"
-					value={password}
-					onChange={handleOnChange}
-				/>
-				<div>
-					<button type="submit" id="confirmation_btn" className="links">
-						Login
-					</button>
-				</div>
-				<span>
-					No account?{" "}
-					<Link className="links" id="signUp-signIn" to={"/signup"}>
-						Sign up
-					</Link>
-				</span>
-			</form>
-			<ToastContainer />
-		</div>
+			<div>
+				<label htmlFor="password" className="input_label">
+					Enter your password
+					<b>*</b>
+				</label>
+			</div>
+			<input
+				className="input_margin"
+				name="password"
+				type="password"
+				placeholder="Password"
+				autoComplete="password"
+				value={password}
+				onChange={handleOnChange}
+			/>
+			<button type="submit" id="confirmation_btn" className="links">
+				Login
+			</button>
+			<span>
+				{"No account? "}
+				<Link className="links" id="signUp-signIn" to={"/signup"}>
+					Sign up
+				</Link>
+			</span>
+		</form>
 	);
 };
