+++ frontend/src/Components/Meeting/DateButtons.js
@@ -1,17 +1,14 @@
 import React, { useEffect, useState } from "react";
-import { useDayViewUpdate } from "../contexts/MeetingContext";
-// import { SERVER_URL } from "../config";
+import { useDayViewUpdate } from "../../contexts/MeetingContext";
 import { v4 as uuidv4 } from "uuid";
-// import { useUpdateUserContext } from "../contexts/LoginContext";
-import APIHandler from "../utils/api-methods";
-
-const api = new APIHandler();
+import APIHandler from "../../utils/api-methods";
 
 const fetchDayMeeting = async (date, monthToDisplay, yearToDisplay) => {
 	try {
 		const dateString = `${String(yearToDisplay).padStart(2, "0")}-${String(
 			monthToDisplay
 		).padStart(2, "0")}-${String(date).padStart(2, "0")}`;
+		const api = new APIHandler();
 		const { data } = await api.GetData(`/meeting/${dateString}`);
 
 		return data;
@@ -19,16 +16,31 @@ const fetchDayMeeting = async (date, monthToDisplay, yearToDisplay) => {
 		console.error(error);
 	}
 };
+const isTodaysDate = (date, month, year) => {
+	const today = new Date();
+	const todaysYear = today.getFullYear();
+	const todaysMonth = (today.getMonth() + 1).toString().padStart(2, "0"); // Months are zero-based
+	const todaysDayNum = today.getDate().toString().padStart(2, "0");
+
+	const customFormattedDate = `${todaysYear}-${todaysMonth}-${todaysDayNum}`;
+	const itemDate = `${year}-${month.toString().padStart(2, "0")}-${date
+		.toString()
+		.padStart(2, "0")}`;
+	if (customFormattedDate === itemDate) return true;
+	else return false;
+};
 
 //Component for a date button
 const DateButtons = ({ date, dayString, month, year, theme }) => {
-	// const { api } = useUpdateUserContext();
 	const { openDayView } = useDayViewUpdate();
 
 	const [isLoading, setIsLoading] = useState(true);
 	const [meetings, setMeetings] = useState();
+	const [isToday, setIsToday] = useState(false);
+
 	useEffect(() => {
 		const renderDayMeetings = async () => {
+			setIsToday(isTodaysDate(date, month, year));
 			try {
 				const meetings = await fetchDayMeeting(date, month, year);
 				setMeetings(meetings);
@@ -39,7 +51,7 @@ const DateButtons = ({ date, dayString, month, year, theme }) => {
 			}
 		};
 		renderDayMeetings();
-	}, [date, month, year]);
+	}, [date, month, year, isToday]);
 
 	return (
 		<div
@@ -48,7 +60,11 @@ const DateButtons = ({ date, dayString, month, year, theme }) => {
 				openDayView(date, dayString, month);
 			}}
 		>
-			{date}
+			<div className="place-center">
+				<div className={`today ${isToday ? "active" : ""}`}>
+					<b>{date}</b>
+				</div>
+			</div>
 			{isLoading ? (
 				<></>
 			) : (
